// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type ContentRelationshipFieldWithData<
  TCustomType extends
    | readonly (prismic.CustomTypeModelFetchCustomTypeLevel1 | string)[]
    | readonly (prismic.CustomTypeModelFetchCustomTypeLevel2 | string)[],
  TLang extends string = string,
> = {
  [ID in Exclude<
    TCustomType[number],
    string
  >['id']]: prismic.ContentRelationshipField<
    ID,
    TLang,
    PickContentRelationshipFieldData<
      Extract<TCustomType[number], { id: ID }>,
      Extract<prismic.Content.AllDocumentTypes, { type: ID }>['data'],
      TLang
    >
  >;
}[Exclude<TCustomType[number], string>['id']];
// ...
type PageDocumentDataSlicesSlice =
  | ProductsGridSlice
  | ProductHeroSlice
  | RelatedContentGridSlice
  | MegaCardsSlice
  | CtaCardSlice
  | MediaGallerySlice
  | ContactInfoWithHoursSlice
  | StoryAndColumnsSlice
  | FeatureGrid3Slice
  | PricingTableSlice
  | HeroBasicSlice
  | FeatureGrid2Slice
  | FeatureGridShowcaseSlice
  | TwoColumnStorySlice
  | TestimonialCarouselSlice
  | TestimonialDetailSlice
  | InfiniteTextMarqueeSlice
  | HeadlineStatsSlice
  | FeatureWithImageGridSlice
  | HeroWithMediaSlice
  | LogoShowcaseSlice
  | HeadlineImageCtaSlice
  | ContactSplitSlice
  | ContactDetailsWithMapSlice
  | MediaFeatureSlice
  | CarouselGallerySlice
  | FaqSplitLayoutSlice
  | FeatureGridSlice
  | ImageCardFeatureSlice
  | HeadlineMediaCallToActionSlice
  | CenteredHeroSlice
  | RichTextSlice;

/**
 * Item in *Page → Sizes*
 */
export interface PageDocumentDataSizesItem {
  /**
   * Size field in *Page → Sizes*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: page.sizes[].size
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  size: prismic.KeyTextField;

  /**
   * Disabled field in *Page → Sizes*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: false
   * - **API ID Path**: page.sizes[].disabled
   * - **Documentation**: https://prismic.io/docs/fields/boolean
   */
  disabled: prismic.BooleanField;
}

/**
 * Content for Page documents
 */
interface PageDocumentData {
  /**
   * Page Title field in *Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Used for breadcrumbs, cards, etc.
   * - **API ID Path**: page.page_title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  page_title: prismic.KeyTextField;

  /**
   * Parent field in *Page*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: page.parent
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/content-relationship
   */
  parent: ContentRelationshipFieldWithData<
    [
      {
        id: 'page';
        fields: [
          {
            id: 'parent';
            customtypes: [{ id: 'page'; fields: ['parent', 'page_title'] }];
          },
          'page_title',
        ];
      },
    ]
  >;

  /**
   * Page Description field in *Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Optional: Used in cards and other summaries
   * - **API ID Path**: page.page_description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  page_description: prismic.KeyTextField;

  /**
   * Featured Image field in *Page*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: page.featured_image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  featured_image: prismic.ImageField<never>;

  /**
   * Tag field in *Page*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: page.tag
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/content-relationship
   */
  tag: ContentRelationshipFieldWithData<
    [{ id: 'tag'; fields: ['tag_name', 'tag_color'] }]
  >;

  /**
   * Publish Date field in *Page*
   *
   * - **Field Type**: Date
   * - **Placeholder**: Optional: Used for articles, etc.
   * - **API ID Path**: page.publish_date
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/date
   */
  publish_date: prismic.DateField;

  /**
   * Slice Zone field in *Page*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: page.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/slices
   */
  slices: prismic.SliceZone<PageDocumentDataSlicesSlice> /**
   * Meta Title field in *Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: page.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: page.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *Page*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: page.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  meta_image: prismic.ImageField<never> /**
   * Stripe Product ID field in *Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: page.stripe_product_id
   * - **Tab**: Product Data
   * - **Documentation**: https://prismic.io/docs/fields/text
   */;
  stripe_product_id: prismic.KeyTextField;

  /**
   * Previous Price field in *Page*
   *
   * - **Field Type**: Number
   * - **Placeholder**: *None*
   * - **API ID Path**: page.previous_price
   * - **Tab**: Product Data
   * - **Documentation**: https://prismic.io/docs/fields/number
   */
  previous_price: prismic.NumberField;

  /**
   * Sizes field in *Page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: page.sizes[]
   * - **Tab**: Product Data
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  sizes: prismic.GroupField<Simplify<PageDocumentDataSizesItem>>;
}

/**
 * Page document from Prismic
 *
 * - **API ID**: `page`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type PageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<Simplify<PageDocumentData>, 'page', Lang>;